+++ common/src/test/java/me/yassigame/sable_beyond/mass/MassFormulaTest.java
@@ -1,0 +1,4 @@
+package me.yassigame.sable_beyond.mass;
+
+public class MassFormulaTest {
+}