+++ CHANGELOG.md
@@ -19,3 +19,8 @@ Example:
 - (Neoforge) Added the ability for Create fans to push sublevels.
 - (Common) Added mass for entities, configurable with formulas, etc.
 - That’s all, I think. Apart from a few things like experimental player interaction / mass, nothing major, and in any case they’re still very much WIP.
+
+## [0.0.2] - Some improvements
+
+- (Neoforge) Added KubeJS support (a doc will be live soon / also i will add more things)
+- (Common) Added a command to have system info (for issue purposes)